+++ tests/utils/samples.py
@@ -1,0 +1,137 @@
+# -*- coding: utf-8 -*-
+
+SAMPLE_DATA = {
+    'bool': [
+        'True',
+        'False',
+    ],
+    'bytearray': [
+        # 'bytearray()',
+        # 'bytearray(1)',
+        # 'bytearray([1, 2, 3])',
+        'bytearray(b"hello world")',
+    ],
+    'bytes': [
+        'b""',
+        'b"This is another string of bytes"',
+        'b"\x01\x75"',  # mixing nonprintable and printable bytes
+    ],
+    'class': [
+        'type(1)',
+        'type("a")',
+        # 'type(object())', # TODO: re-enable this when object() is implemented
+        'type("MyClass", (object,), {})',
+    ],
+    'complex': [
+        '0j',
+        '1j',
+        '0.7071067811865476+0.7071067811865475j',  # sqrt(0.5)*(1+1j) -- magnitude 1, but not pure
+        '3.14159265j',
+        '-5j',
+        '1+2j',
+        '3-4j',
+        '-2+0j',
+        # '179769313486231590772930519078902473361797697894230657273430081157732675805500963132708477322407536021120113879871393357658789768814416622492847430639474124377767893424865485276302219601246094119453082952085005768838150682342462881473913110540827237163350510684586298239947245938479716304835356329624224137216+100j',  # NOQA
+    ],
+    'dict': [
+        '{}',
+        '{"a": 1, "c": 2.3456, "d": "another"}',
+        '{"a": {"a": 1}, "b": {"b": 2}}',
+    ],
+    'float': [
+        '2.3456',
+        '0.0',
+        '-3.14159',
+        '-4.81756',
+        '5.5',
+        '-3.5',
+        '4.5',
+        '-4.5',
+    ],
+    'frozenset': [
+        'frozenset()',
+        'frozenset([1])',
+        # 'frozenset({"1"})', this reveals some bugs in our code
+        'frozenset({1, 2.3456, "another"})',
+    ],
+    'int': [
+        '3',
+        '0',
+        '-5',
+        '-3',
+        '5',
+        '1',
+        '-1',
+        '9223372036854775807',
+        '9223372036854775808',
+        '-9223372036854775807',
+        '-9223372036854775808',
+        '18446744073709551615',
+        '18446744073709551616',
+        '18446744073709551617',
+        '-18446744073709551615',
+        '-18446744073709551616',
+        '-18446744073709551617',
+        '1361129467683753853853498429727072845824',
+        '-1361129467683753853853498429727072845824',
+        '179769313486231590772930519078902473361797697894230657273430081157732675805500963132708477322407536021120113879871393357658789768814416622492847430639474124377767893424865485276302219601246094119453082952085005768838150682342462881473913110540827237163350510684586298239947245938479716304835356329624224137216',  # NOQA
+        '-179769313486231590772930519078902473361797697894230657273430081157732675805500963132708477322407536021120113879871393357658789768814416622492847430639474124377767893424865485276302219601246094119453082952085005768838150682342462881473913110540827237163350510684586298239947245938479716304835356329624224137216',  # NOQA
+    ],
+    'list': [
+        '[]',
+        '[1]',
+        '[3, 4, 5]',
+        '[1, 2, 3, 4, 5, 6, 7, 8, 9, 10]',
+        '["a","b","c"]',
+        '[[1, 2], [3, 4]]',
+    ],
+    'range': [
+        'range(0)',
+        'range(5)',
+        'range(2, 7)',
+        'range(2, 7, 2)',
+        'range(7, 2, -1)',
+        'range(7, 2, -2)',
+    ],
+    'set': [
+        'set()',
+        '{1, 2.3456, "another"}',
+    ],
+    'slice': [
+        'slice(0)',
+        'slice(5)',
+        'slice(2, 7)',
+        'slice(2, 7, 2)',
+        'slice(7, 2, -1)',
+        'slice(7, 2, -2)',
+    ],
+    'str': [
+        '""',
+        '"3"',
+        '"This is another string"',
+        '"Mÿ hôvèrçràft îß fûłl öf éêlś"',
+        '"/// One arg: %s"',
+        '"Three args: %s | %s | %s"',
+    ],
+    'tuple': [
+        '()',
+        '(False,)',
+        '(1,)',
+        '(1, 2)',
+        '(1, 2, 3, 4, 5, 6, 7, 8, 9, 10)',
+        '(3, 1.2, True, )',
+        '(1, 2.3456, "another")',
+        '((1, 2), (3, 4))',
+        '((3, 4), (1, 2))'
+    ],
+    'None': [
+        'None',
+    ],
+    'NotImplemented': [
+        'NotImplemented',
+    ],
+}
+
+SAMPLE_SUBSTITUTIONS = {
+
+}
